backend/requirements.txt
boto3>=1.28.0
requests>=2.31.0
pystac-client>=0.7.0
stackstac>=0.4.0
xarray>=2023.8.0
rioxarray>=0.16.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.40.0
folium>=0.15.0
streamlit-folium>=0.20.0
reportlab>=4.0.0

